docker/docker_db/pyproject.toml
[tool.poetry]
name = "docker-db"
version = "0.1.0"
description = ""
authors = ["Muhammad Saad <saadjameel2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlmodel = "^0.0.19"
psycopg2-binary = "^2.9.9"
pydantic = "^2.7.3"
sqlalchemy = "^2.0.30"
databases = "^0.9.0"
asyncpg = "^0.29.0"
psycopg = "^3.1.19"
starlette = "^0.37.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
